% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridutils-package.R
\docType{package}
\name{gridutils-package}
\alias{gridutils}
\alias{gridutils-package}
\title{gridutils: Small Utilities For grid Package}
\description{
(This is my personal practice to play with grid).
}
\author{
\strong{Maintainer}: Hiroaki Yutani \email{yutani.ini@gamil.com}

}
\keyword{internal}
