% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewport.R
\name{you-are-here}
\alias{you-are-here}
\alias{gu_you_are_here}
\title{Highlight the current viewport}
\usage{
gu_you_are_here()
}
\description{
Unlike \code{\link[grid:showViewport]{grid::showViewport()}}, \code{gu_show_vp()} soon remove the highlight after
a second
}
