% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridutils-package.R
\docType{package}
\name{gridutils-package}
\alias{gridutils}
\alias{gridutils-package}
\title{gridutils: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Hiroaki Yutani \email{yutani.ini@gamil.com}

}
\keyword{internal}
